+++ src/mantle/testing/concerns/trait-element-assertions.php
@@ -2,6 +2,8 @@
 /**
  * Element_Assertions trait file
  *
+ * phpcs:disable Squiz.Commenting.FunctionComment.ParamNameNoMatch, Squiz.Commenting.FunctionComment.MissingParamTag
+ *
  * @package Mantle
  */
 
@@ -9,6 +11,7 @@ namespace Mantle\Testing\Concerns;
 
 use PHPUnit\Framework\Assert as PHPUnit;
 use DOMDocument;
+use DOMNode;
 use DOMXPath;
 use Symfony\Component\CssSelector\CssSelectorConverter;
 
@@ -52,11 +55,15 @@ trait Element_Assertions {
 	 * Assert that an element exists in the response.
 	 *
 	 * @param string $expression The XPath expression to execute.
+	 * @param string $message Optional message to display on failure.
 	 */
-	public function assertElementExists( string $expression ): static {
+	public function assertElementExists( string $expression, string $message = null ): static {
 		$nodes = ( new DOMXPath( $this->get_dom_document() ) )->query( $expression );
 
-		PHPUnit::assertTrue( ! $nodes ? false : $nodes->length > 0 );
+		PHPUnit::assertTrue(
+			! $nodes ? false : $nodes->length > 0,
+			$message ?? 'Element not found for expression: ' . $expression,
+		);
 
 		return $this;
 	}
@@ -71,7 +78,7 @@ trait Element_Assertions {
 			$id = substr( $id, 1 );
 		}
 
-		return $this->assertElementExists( sprintf( '//*[@id="%s"]', $id ) );
+		return $this->assertElementExists( sprintf( '//*[@id="%s"]', $id ), "Element not found for ID: $id" );
 	}
 
 	/**
@@ -84,18 +91,25 @@ trait Element_Assertions {
 			$classname = substr( $classname, 1 );
 		}
 
-		return $this->assertElementExists( sprintf( '//*[contains(concat(" ", normalize-space(@class), " "), " %s ")]', $classname ) );
+		return $this->assertElementExists(
+			sprintf( '//*[contains(concat(" ", normalize-space(@class), " "), " %s ")]', $classname ),
+			"Element not found for class: $classname"
+		);
 	}
 
 	/**
 	 * Assert that an element is missing in the response.
 	 *
 	 * @param string $expression The XPath expression to execute.
+	 * @param string $message    The message to display if the assertion fails.
 	 */
-	public function assertElementMissing( string $expression ): static {
+	public function assertElementMissing( string $expression, string $message = null ): static {
 		$nodes = ( new DOMXPath( $this->get_dom_document() ) )->query( $expression );
 
-		PHPUnit::assertTrue( false === $nodes || 0 === $nodes->length );
+		PHPUnit::assertTrue(
+			false === $nodes || 0 === $nodes->length,
+			$message ?? "Element found for expression: $expression"
+		);
 
 		return $this;
 	}
@@ -110,7 +124,7 @@ trait Element_Assertions {
 			$id = substr( $id, 1 );
 		}
 
-		return $this->assertElementMissing( sprintf( '//*[@id="%s"]', $id ) );
+		return $this->assertElementMissing( sprintf( '//*[@id="%s"]', $id ), "Element found for ID: $id" );
 	}
 
 	/**
@@ -132,7 +146,7 @@ trait Element_Assertions {
 	 * @param string $type The type of element to check.
 	 */
 	public function assertElementExistsByTagName( string $type ): static {
-		return $this->assertElementExists( sprintf( '//*[local-name()="%s"]', $type ) );
+		return $this->assertElementExists( sprintf( '//*[local-name()="%s"]', $type ), "Element not found for tag: $type" );
 	}
 
 	/**
@@ -141,7 +155,7 @@ trait Element_Assertions {
 	 * @param string $type The type of element to check.
 	 */
 	public function assertElementMissingByTagName( string $type ): static {
-		return $this->assertElementMissing( sprintf( '//*[local-name()="%s"]', $type ) );
+		return $this->assertElementMissing( sprintf( '//*[local-name()="%s"]', $type ), "Element found for tag: $type" );
 	}
 
 	/**
@@ -150,7 +164,7 @@ trait Element_Assertions {
 	 * @param string $selector The selector to use.
 	 */
 	public function assertElementExistsByQuerySelector( string $selector ): static {
-		return $this->assertElementExists( $this->convert_query_selector( $selector ) );
+		return $this->assertElementExists( $this->convert_query_selector( $selector ), "Element not found for selector: $selector" );
 	}
 
 	/**
@@ -168,7 +182,7 @@ trait Element_Assertions {
 	 * @param string $selector The selector to use.
 	 */
 	public function assertElementMissingByQuerySelector( string $selector ): static {
-		return $this->assertElementMissing( $this->convert_query_selector( $selector ) );
+		return $this->assertElementMissing( $this->convert_query_selector( $selector ), "Element found for selector: $selector" );
 	}
 
 	/**
@@ -178,5 +192,92 @@ trait Element_Assertions {
 	 */
 	public function assertQuerySelectorMissing( string $selector ): static {
 		return $this->assertElementMissingByQuerySelector( $selector );
+	}
+
+	/**
+	 * Assert against the expected number of elements for an expression.
+	 *
+	 * @param string $expression The XPath expression to execute.
+	 * @param int    $expected The expected number of elements.
+	 */
+	public function assertElementCount( string $expression, int $expected ): static {
+		$nodes = ( new DOMXPath( $this->get_dom_document() ) )->query( $expression );
+
+		PHPUnit::assertEquals( $expected, $nodes->length, 'Unexpected number of elements found.' );
+
+		return $this;
+	}
+
+	/**
+	 * Assert against the expected number of elements for a query selector.
+	 *
+	 * @param string $selector The selector to use.
+	 * @param int    $expected The expected number of elements.
+	 */
+	public function assertQuerySelectorCount( string $selector, int $expected ): static {
+		return $this->assertElementCount( $this->convert_query_selector( $selector ), $expected );
+	}
+
+	/**
+	 * Assert an element exists by test ID.
+	 *
+	 * @param string $test_id The test ID to check.
+	 */
+	public function assertElementExistsByTestId( string $test_id ): static {
+		return $this->assertQuerySelectorExists( "[data-testid=\"$test_id\"]" );
+	}
+
+	/**
+	 * Assert an element is missing by test ID.
+	 *
+	 * @param string $test_id The test ID to check.
+	 */
+	public function assertElementMissingByTestId( string $test_id ): static {
+		return $this->assertQuerySelectorMissing( "[data-testid=\"$test_id\"]" );
+	}
+
+	/**
+	 * Assert that an element passes a custom assertion.
+	 *
+	 * The assertion will be called for each node found by the expression.
+	 *
+	 * @param string           $expression The XPath expression to execute.
+	 * @param callable(DOMNode $node): bool $assertion The assertion to run.
+	 * @param bool             $pass_any Pass if any of the nodes pass the assertion. Otherwise, all must pass.
+	 */
+	public function assertElement( string $expression, callable $assertion, bool $pass_any = false ): static {
+		$nodes = ( new DOMXPath( $this->get_dom_document() ) )->query( $expression );
+
+		if ( ! $nodes ) {
+			PHPUnit::fail( 'No nodes found for expression: ' . $expression );
+		}
+
+		foreach ( $nodes as $node ) {
+			if ( $assertion( $node ) ) {
+				// If we're passing on any, we can return early.
+				if ( $pass_any ) {
+					return $this;
+				}
+			} elseif ( ! $pass_any ) {
+				PHPUnit::fail( 'Assertion failed for node: ' . $node->nodeName ); // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
+			}
+		}
+
+		PHPUnit::assertTrue( true, 'All nodes passed assertion.' );
+
+		return $this;
+	}
+
+	/**
+	 * Assert that an element passes a custom assertion by query selector.
+	 *
+	 * The assertion will be called for each node found by the selector.
+	 *
+	 * @param string           $selector The selector to use.
+	 * @param callable(DOMNode $node): bool $assertion The assertion to run.
+	 * @param bool             $pass_any Pass if any of the nodes pass the assertion. Otherwise, all must pass.
+	 */
+	public function assertQuerySelector( string $selector, callable $assertion, bool $pass_any = false ): static {
+		return $this->assertElement( $this->convert_query_selector( $selector ), $assertion, $pass_any );
 	}
 }