+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v1.0.2 - 2024-04-10
+## v1.0.2 - 2024-04-15
+
+### Added
+
+- Added `html_string()` helper to make assertions against a HTML string easier in testing.
+- Added new assertion methods to test against elements.
 
 ### Fixed